+++ src/node_query_system/search_criteria/custom.cpp
@@ -12,7 +12,7 @@ UtilityAICustomSearchCriterion::~UtilityAICustomSearchCriterion() {
 
 }
 
-void UtilityAICustomSearchCriterion::apply_criterion( Node* node, bool& filter_out, double& score ) {
+void UtilityAICustomSearchCriterion::apply_criterion( Node* node, bool& filter_out, float& score ) {
     if( !has_method("apply_criterion")) return;
     call("apply_criterion", node);
     filter_out = _is_filtered;
